analytics/terraform/spark-k8s-operator/examples/benchmark/scale/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
coverage==7.6.10
durationpy==0.9
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
google-auth==2.38.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kubernetes==31.0.0
kubernetes-client==0.1.8
locust==2.32.6
MarkupSafe==3.0.2
msgpack==1.1.0
mypy==1.14.1
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
table-logger==0.3.6
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
